objectDetection/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
